seal

+++ src/Files/Metadata/Exif/Thumbnail.php
@@ -16,8 +16,8 @@ use ImageKit\Core\Contracts\BaseModel;
  *   resolutionUnit?: int|null,
  *   thumbnailLength?: int|null,
  *   thumbnailOffset?: int|null,
- *   xResolution?: int|null,
- *   yResolution?: int|null,
+ *   xResolution?: float|null,
+ *   yResolution?: float|null,
  * }
  */
 final class Thumbnail implements BaseModel
@@ -38,10 +38,10 @@ final class Thumbnail implements BaseModel
     public ?int $thumbnailOffset;
 
     #[Optional('XResolution')]
-    public ?int $xResolution;
+    public ?float $xResolution;
 
     #[Optional('YResolution')]
-    public ?int $yResolution;
+    public ?float $yResolution;
 
     public function __construct()
     {
@@ -58,8 +58,8 @@ final class Thumbnail implements BaseModel
         ?int $resolutionUnit = null,
         ?int $thumbnailLength = null,
         ?int $thumbnailOffset = null,
-        ?int $xResolution = null,
-        ?int $yResolution = null,
+        ?float $xResolution = null,
+        ?float $yResolution = null,
     ): self {
         $self = new self;
 
@@ -105,7 +105,7 @@ final class Thumbnail implements BaseModel
         return $self;
     }
 
-    public function withXResolution(int $xResolution): self
+    public function withXResolution(float $xResolution): self
     {
         $self = clone $this;
         $self['xResolution'] = $xResolution;
@@ -113,7 +113,7 @@ final class Thumbnail implements BaseModel
         return $self;
     }
 
-    public function withYResolution(int $yResolution): self
+    public function withYResolution(float $yResolution): self
     {
         $self = clone $this;
         $self['yResolution'] = $yResolution;
